requirements.txt
django==5.0.1
PyMySQL==1.1.0
supabase==1.0.3
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
python-magic==0.4.27
dj-database-url==2.1.0
whitenoise==6.5.0
gunicorn==21.2.0
setuptools==69.0.2